% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Center}
\usage{
center(x)
}
\arguments{
\item{x}{a vector of a numeric or integer values}
}
\value{
a vector of a numeric
}
\description{
Center returns a centered copy of a vector of values, it subtracts the mean
of the vector from the value.
}
\examples{
a<- 50:100
center(a)

}
